separate inner classes
