+++ app/src/main/java/com/nextcloud/ui/tags/TagManagementViewModel.kt
@@ -10,6 +10,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.nextcloud.client.account.CurrentAccountProvider
 import com.nextcloud.client.network.ClientFactory
+import com.nextcloud.ui.tags.model.TagUiState
 import com.owncloud.android.lib.resources.tags.CreateTagRemoteOperation
 import com.owncloud.android.lib.resources.tags.DeleteTagRemoteOperation
 import com.owncloud.android.lib.resources.tags.GetTagsRemoteOperation
@@ -26,13 +27,6 @@ class TagManagementViewModel @Inject constructor(
     private val clientFactory: ClientFactory,
     private val currentAccountProvider: CurrentAccountProvider
 ) : ViewModel() {
-
-    sealed interface TagUiState {
-        object Loading : TagUiState
-        data class Loaded(val allTags: List<Tag>, val assignedTagIds: Set<String>, val query: String = "") : TagUiState
-
-        data class Error(val message: String) : TagUiState
-    }
 
     private val _uiState = MutableStateFlow<TagUiState>(TagUiState.Loading)
     val uiState: StateFlow<TagUiState> = _uiState
